requirements.txt
affine==2.3.1
anyio==3.6.2
attrs==22.1.0
black==22.10.0
certifi==2022.9.24
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.0.6
cycler==0.11.0
fastapi==0.86.0
fonttools==4.38.0
h11==0.14.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.2
mypy-extensions==0.4.3
numpy==1.23.4
packaging==21.3
pathspec==0.10.1
Pillow==9.3.0
platformdirs==2.5.3
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
rasterio==1.3.3
six==1.16.0
sniffio==1.3.0
snuggs==1.4.7
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.19.0
